Convert strip unused keys script
