+++ packages/cli/package.json
@@ -30,8 +30,7 @@
     "validate-last-values": "yarn workspace @corona-dashboard/common build && ts-node src/scripts/validate-last-values.ts",
     "validate-features": "yarn workspace @corona-dashboard/common build && ts-node src/scripts/validate-features.ts",
     "generate-typescript": "yarn workspace @corona-dashboard/common build && ts-node src/scripts/generate-data-types.ts",
-    "generate-sitemap": "ts-node src/scripts/generate-sitemap.ts",
-    "strip-unused-locale-keys": "yarn workspace @corona-dashboard/cms lokalize:export --clean-json && ts-node src/locale/strip-unused-locale-keys.ts && yarn workspace @corona-dashboard/cms lokalize:export --clean-json && yarn workspace @corona-dashboard/app compile"
+    "generate-sitemap": "ts-node src/scripts/generate-sitemap.ts"
   },
   "devDependencies": {
     "@types/meow": "^5.0.0"
